Add files via upload
The commom positive keyboard is like a tic-tac-toe field (3 rows and 3 columns) -- press the middle button - the letter "O" comes out. Why, because the rest of the buttons are shaped like this letter. We press the middle lower button (in 3*3 keypad) and the middle one - we get "n". Two 3*3 keypads make one hybrid positive keyboard. So, by pressing on the qwerty-keyboard in our editor "S"-letter - results "o", and by pressing "S"-letter and "X"-letter -- "n".